Baselines/src/requirements.txt
pandas
scikit_learn
auto-sklearn
h2o
xgboost